refactor(theme): type themeLabel against ThemeName
themeLabel() took a bare string and re-encoded the special theme names,
so a misspelling rendered as Apple Music and hid the typo in the tray.
ThemeName and isThemeName() move beside the palettes, and both label
lookups now use the existing map instead of a linear find.
